% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-hard-fast.R
\name{compute_adj_term}
\alias{compute_adj_term}
\title{Adjustment Term for qhat Uncertainty}
\usage{
compute_adj_term(Hmat, var.qhat, Lm1, H)
}
\arguments{
\item{Hmat}{An n x (Lm1 * H) basis matrix (e.g. from \code{build_H}).}

\item{var.qhat}{An n x Lm1 matrix of pointwise variances of qhat
(i.e. \code{qhat.se^2}).}

\item{Lm1}{Integer. Number of non-reference groups (L - 1).}

\item{H}{Integer. Number of basis columns per non-reference group.}
}
\value{
A (Lm1 * H) x (Lm1 * H) symmetric matrix representing the
qhat-uncertainty adjustment.
}
\description{
Computes the block-diagonal adjustment matrix that accounts for
estimation uncertainty in the external predicted probabilities
(\code{qhat}). Used to inflate the variance of the t-block in
sandwich standard errors.
}
\keyword{internal}
