% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{conf.est.thm3}
\alias{conf.est.thm3}
\title{Theorem 3 standard errors}
\usage{
conf.est.thm3(par, X, y, qhat, qhat.se, Hmat, A)
}
\arguments{
\item{par}{Hard-format parameter vector (beta, Theta, alpha, tmat).}

\item{X}{Design matrix (n x p).}

\item{y}{Integer response vector of length n (class labels 1..K).}

\item{qhat}{External probability matrix (n x (L-1)).}

\item{qhat.se}{Standard errors of external probabilities (n x (L-1)).}

\item{Hmat}{Basis matrix for EL constraints.}

\item{A}{Group matrix (K x L) from \code{group_matrix()}.}
}
\value{
A list with components:
\describe{
\item{se.thm3}{Unadjusted Theorem 3 SE vector.}
\item{se.thm3.adj}{Adjusted Theorem 3 SE vector.}
\item{Sigma_theta}{Unadjusted theta-block covariance.}
\item{Sigma_theta_adj}{Adjusted theta-block covariance.}
\item{I_theta_inv}{Inverse Fisher information for theta.}
}
}
\description{
Computes standard errors using Theorem 3 of the fused multinomial
model, evaluating the Hessian at t=0 (unfused model).
}
