% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{gradient_hard}
\alias{gradient_hard}
\title{Gradient}
\usage{
gradient_hard(par, X, y, qhat, Hmat, A, tau_tmat = 1)
}
\arguments{
\item{par}{Packed hard-mode parameter vector.}

\item{X}{n x p covariate matrix.}

\item{y}{Integer response vector.}

\item{qhat}{n x (L-1) external probability matrix.}

\item{Hmat}{Basis matrix.}

\item{A}{K x L group matrix.}

\item{tau_tmat}{L2 penalty on tmat (default 1).}
}
\value{
Numeric gradient vector of length Dnoq.
}
\description{
Analytic gradient of the fused negative log-likelihood.
}
