% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_matrix}
\alias{group_matrix}
\title{Group Membership Matrix}
\usage{
group_matrix(groups, K)
}
\arguments{
\item{groups}{A list of length L, where each element is an integer vector
of class indices belonging to that group.}

\item{K}{Integer. Total number of classes.}
}
\value{
A K x L binary matrix where \code{A[k, l] = 1} if class k belongs
to group l, and 0 otherwise. Each class belongs to exactly one group.
}
\description{
Creates a K x L binary partition matrix from a list of group assignments.
}
