% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{make_hard_indices}
\alias{make_hard_indices}
\title{Parameter Index Ranges}
\usage{
make_hard_indices(d, p, Lm1, H)
}
\arguments{
\item{d}{Number of non-baseline classes (K - 1).}

\item{p}{Number of covariates.}

\item{Lm1}{Number of non-reference groups (L - 1).}

\item{H}{Basis columns per group.}
}
\value{
Named list of index vectors and block sizes.
}
\description{
Parameter Index Ranges
}
\keyword{internal}
