% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective-hard-fast.R
\name{objective.hard}
\alias{objective.hard}
\title{Hard-Mode Objective Value}
\usage{
objective.hard(par, X, y, qhat, Hmat, A)
}
\arguments{
\item{par}{Packed hard-mode parameter vector (see \code{\link{pack.hard}}).}

\item{X}{An n x p covariate matrix.}

\item{y}{An integer vector of length n with class labels in 1, ..., K.}

\item{qhat}{An n x (L-1) matrix of external predicted group
probabilities for non-reference groups.}

\item{Hmat}{An n x ((L-1)*H) basis matrix for the EL constraint.}

\item{A}{A K x L group membership matrix (see \code{\link{group_matrix}}).}
}
\value{
A list with elements:
\describe{
\item{obj}{Scalar. The negative log-likelihood value.}
\item{P}{An n x K matrix of primary softmax probabilities.}
\item{Pext_L}{An n x L matrix of external group probabilities.}
}
}
\description{
Computes the negative log-likelihood of the fused multinomial model
with empirical-likelihood constraint, evaluated at hard-mode
parameters (q fixed at qhat).
}
