% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softmax_first}
\alias{softmax_first}
\title{Baseline-Constrained Softmax}
\usage{
softmax_first(eta_mat)
}
\arguments{
\item{eta_mat}{An n x (K-1) matrix of linear predictors for classes
2, ..., K.}
}
\value{
An n x K probability matrix. Column 1 is the baseline class.
Rows sum to 1.
}
\description{
Computes softmax probabilities with class 1 as the reference category
(linear predictor fixed at 0).
}
