% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{build_Hmat_interleaved}
\alias{build_Hmat_interleaved}
\title{Interleave Basis Matrix for Multiple Groups}
\usage{
build_Hmat_interleaved(H_base, Lm1)
}
\arguments{
\item{H_base}{An n x H matrix (single-group basis).}

\item{Lm1}{Integer. Number of non-reference groups (L - 1).}
}
\value{
An n x (H * Lm1) interleaved matrix.
}
\description{
Replicates a single-group basis matrix into an interleaved layout
for Lm1 non-reference groups. Column ordering: for basis k and
group m, output column = m + (k-1)*Lm1.
}
